# GraphWebApi/wwwroot/OpenApi.yaml
openapi: 3.0.0
info: 
  title: "Graph Explorer API" 
  version: "1.0.0" 
servers:
  - url: https://graphexplorerapi.azurewebsites.net/
  - url: https://localhost:44399/
paths:
...
  /samples/{id}:
    get:
      tags:
        - samples
      description: Get a specific sample query by its Id.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Ok
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/sample'
        204:
          description: No content
        404:
          description: Not found
        500:
          description: Internal server error
    put:
      tags:
        - samples
      description: Update a specific sample query by its Id.
      security:
        - OAuth2Implicit: [api://xxxx-xxxx-xxxx-xxxx-xxxx/SampleQueries.ReadWriteAll]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/sample'
            example:
              {
                "category": "Security",
                "method": "GET",
                "humanName": "alerts filter by destination address",
                "requestUrl": "/v1.0/security/alerts?$filter=networkConnections/any(d:d/destinationAddress eq '{destination-address}')",
                "docLink": "https://developer.microsoft.com/en-us/graph/docs/api-reference/v1.0/api/alert_list",
                "tip": "This query requires a destination address. Run https://graph.microsoft.com/v1.0/security/alerts?$top=1 and search the results for a destinationAddress property.",
                "skipTest": false
              }
      responses:
        200:
          description: Ok
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/sample'
        401:
          description: Unauthorized
        403:
          description: Forbidden
        404:
          description: Not found
        500:
          description: Internal server error
    delete:
      tags:
        - samples
      description: Delete a specific sample query by its Id.
      security:
        - OAuth2Implicit: [api://xxxx-xxxx-xxxx-xxxx-xxxx/SampleQueries.ReadWriteAll]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        204:
          description: Deleted successfully
        401:
          description: Unauthorized
        403:
          description: Forbidden
        404:
          description: Not found
        500:
          description: Internal server error
...
  /samplesadmin:
    get:
      tags:
        - admin
      description: Gets all category policies or filter by userPrincipalName or categoryName.
      parameters:
        - name: userPrincipalName
          in: query
          required: false
          schema:
            type: string
        - name: categoryName
          in: query
          required: false
          schema:
            type: string
      responses:
        200:
          description: List of sample queries category policies.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/categoryPolicy'
              example: 
                {
                  "CategoryPolicies": [
                    {
                      "categoryName": "Getting Started",
                      "userClaims": [
                        {
                          "userName": "john doe",
                          "userPrincipalName": "john.doe@microsoft.com",
                          "userPermissions": [
                            {
                              "name": "POST",
                              "value": true
                            },
                            {
                              "name": "PUT",
                              "value": true
                            },
                            {
                              "name": "DELETE",
                              "value": false
                            }
                          ]
                        },
                        {
                          "userName": "jane doe",
                          "userPrincipalName": "jane.doe@microsoft.com",
                          "userPermissions": [
                            {
                              "name": "POST",
                              "value": true
                            },
                            {
                              "name": "PUT",
                              "value": true
                            },
                            {
                              "name": "DELETE",
                              "value": true
                            }
                          ]
                        }
                      ]
                    }
                  ]
                }
                
        404:
          description: Not found
        500:
          description: Internal server error
    post:
      tags:
        - admin
      description: Adds a new user claim to a category policy.
      security:
        - OAuth2Implicit: [api://xxxx-xxxx-xxxx-xxxx-xxxx/SampleQueries.ReadWriteAll]
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/categoryPolicy'
            example:
              {
                "categoryName": "Getting Started",
                "userClaims": [
                  {
                    "userName": "john doe",
                    "userPrincipalName": "john.doe@microsoft.com",
                    "userPermissions": [
                      {
                        "name": "POST",
                        "value": true
                      },
                      {
                        "name": "PUT",
                        "value": true
                      },
                      {
                        "name": "DELETE",
                        "value": true
                      }
                    ]
                  }
                ]
              }
      responses:
        201:
          description: Returns an array of all the user claims along with the newly created claim in the category policy.
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/categoryPolicy'
        401:
          description: Unauthorized
        403:
          description: Forbidden
        500:
          description: Internal server error
    put:
      tags:
        - admin
      description: Updates a user claim in a category policy.
      security:
        - OAuth2Implicit: [api://xxxx-xxxx-xxxx-xxxx-xxxx/SampleQueries.ReadWriteAll]
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/categoryPolicy'
            example: 
              {
                "categoryName": "Getting Started",
                "userClaims": [
                  {
                    "userName": "john doe",
                    "userPrincipalName": "john.doe@microsoft.com",
                    "userPermissions": [
                      {
                        "name": "POST",
                        "value": true
                      },
                      {
                        "name": "PUT",
                        "value": true
                      },
                      {
                        "name": "DELETE",
                        "value": false
                      }
                    ]
                  }
                ]
              }
      responses:
        200:
          description: Ok
          content:
              'application/json':
                schema:
                  $ref: '#/components/schemas/categoryPolicy'
        400:
          description: Bad request
        401:
          description: Unauthorized
        403:
          description: Forbidden
        404:
          description: Not found
        500:
          description: Internal server error
    delete:
      tags:
        - admin
      description: Deletes a user claim in a category policy.
      security:
        - OAuth2Implicit: [api://xxxx-xxxx-xxxx-xxxx-xxxx/SampleQueries.ReadWriteAll]
      parameters:
        - name: userPrincipalName
          in: query
          required: true
          schema:
            type: string
        - name: categoryName
          in: query
          required: true
          schema:
            type: string
      responses:
        204:
          description: Deleted successfully
        401:
          description: Unauthorized
        403:
          description: Forbidden
        404:
          description: Not found
        500:
          description: Internal server error
components:
  schemas:
    sample:
      type: object
      required:
        - category
        - method
        - humanName
        - requestUrl
      properties:
        id:
          type: string
        category:
          type: string
        method:
          type: string
          enum:
            - GET
            - PUT
            - POST
            - DELETE
            - PATCH
        humanName:
          type: string
        requestUrl:
          type: string
        docLink:
          type: string
        tip:
          type: string
        skipTest: 
          type: boolean
        postBody:
          type: string
        headers:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              value:
                type: string
    categoryPolicy:
      type: object
      properties:
        categoryName:
          type: string
        userClaims:
          type: array
          items:
            type: object
            properties:
              userName:
                type: string
              userPrincipalName:
                type: string
              userPermissions:
                type: array
                items:
                  type: object
                  properties:
                    name:
                      type: string
                    value:
                      type: boolean
...
  securitySchemes:
    OAuth2Implicit:
      type: oauth2
      description: This API uses OAuth 2 with the implicit grant flow to authenticate with Azure AD.
      flows:
        implicit:
          authorizationUrl: https://login.microsoftonline.com/common/oauth2/v2.0/authorize
          scopes: 
            api://xxxx-xxxx-xxxx-xxxx-xxxx/SampleQueries.ReadWriteAll: writes, updates and deletes workload samples.